refactor: extract test setup logic into helper
